DUOS-512 Move models out into separate files for more readability
