+++ automation/src/test/scala/org/broadinstitute/dsp/consent/models/DataUseModels.scala
@@ -1,0 +1,11 @@
+package  org.broadinstitute.dsp.consent.models
+
+object DataUseModels {
+    case class DataUse(
+        hmbResearch: Option[Boolean],
+        populationOriginsAncestry: Option[Boolean],
+        commercialUse: Option[Boolean],
+        pediatric: Option[Boolean],
+        collaboratorRequired: Option[Boolean]
+    )
+}
