+++ automation/src/test/scala/org/broadinstitute/dsp/consent/models/JsonProtocols.scala
@@ -2,328 +2,17 @@ package org.broadinstitute.dsp.consent.models
 
 import spray.json._
 import DefaultJsonProtocol._
+import org.broadinstitute.dsp.consent.models.UserModels._
+import org.broadinstitute.dsp.consent.models.WhiteListModels._
+import org.broadinstitute.dsp.consent.models.ResearcherModels._
+import org.broadinstitute.dsp.consent.models.DataSetModels._
+import org.broadinstitute.dsp.consent.models.DataUseModels._
+import org.broadinstitute.dsp.consent.models.DataAccessRequestModels._
+import org.broadinstitute.dsp.consent.models.NihModels._
+import org.broadinstitute.dsp.consent.models.DacModels._
+
 
 import scala.util.{Failure, Success, Try}
-
-case class ResearcherProperty(
-    propertyId: Int,
-    userId: Int,
-    propertyKey: String,
-    propertyValue: String
-)
-
-case class ResearcherInfo(
-    profileName: Option[String] = None,
-    academicEmail: Option[String] = None,
-    institution: Option[String] = None,
-    department: Option[String] = None,
-    address1: Option[String] = None,
-    city: Option[String] = None,
-    zipcode: Option[String] = None,
-    isThePI: Option[String] = None,
-    country: Option[String] = None,
-    division: Option[String] = None,
-    address2: Option[String] = None,
-    state: Option[String] = None,
-    eRACommonsID: Option[String] = None,
-    pubmedID: Option[String] = None,
-    scientificURL: Option[String] = None,
-    havePI: Option[String] = None,
-    piName: Option[String] = None,
-    piEmail: Option[String] = None,
-    piERACommonsID: Option[String] = None,
-    completed: Option[String] = None,
-    investigator: Option[String] = None,
-    eraExpiration: Option[String] = None,
-    eraAuthorized: Option[String] = None,
-    nihUsername: Option[String] = None,
-    linkedIn: Option[String] = None,
-    researcherGate: Option[String] = None,
-    orcid: Option[String] = None,
-    checkNotifications: Option[String] = None,
-    libraryCards: Option[Seq[String]] = None,
-    libraryCardEntries: Option[Seq[WhiteListEntry]] = None
-)
-
-case class UserRole(
-    userRoleId: Int,
-    userId: Int,
-    roleId: Int,
-    name: String,
-    dacId: Option[Int]
-)
-
-case class WhiteListEntry(
-    organization: String,
-    commonsId: String,
-    name: String,
-    email: String,
-    signingOfficialName: String,
-    signingOfficialEmail: String,
-    itDirectorName: String,
-    itDirectorEmail: String
-)
-
-case class User(
-    dacUserId: Int, 
-    email: String,
-    displayName: String,
-    createDate: String,
-    roles: Seq[UserRole],
-    emailPreference: Boolean,
-    status: String,
-    researcherProperties: Seq[ResearcherProperty],
-    whitelistEntries: Seq[WhiteListEntry]
-)
-
-case class DataSet(
-    dacId: Int,
-    dataSetId: Int,
-    consentId: String,
-    translatedUseRestriction: Option[String],
-    deletable: Option[Boolean],
-    properties: Seq[DataSetProperty],
-    active: Boolean,
-    needsApproval: Boolean,
-    isAssociatedToDataOwners: Option[Boolean],
-    //updateAssociationToDataOwnerAllowed: Option[Boolean],
-    alias: String,
-    objectId: Option[String],
-    createDate: Option[Long],
-    createUserId: Option[Int],
-    updateDate: Option[Long],
-    updateUserId: Option[Int],
-    dataUse: DataUse
-)
-
-case class DataUse(
-    hmbResearch: Option[Boolean],
-    populationOriginsAncestry: Option[Boolean],
-    commercialUse: Option[Boolean],
-    pediatric: Option[Boolean],
-    collaboratorRequired: Option[Boolean]
-)
-
-case class DataSetProperty(
-    propertyName: String,
-    propertyValue: String
-)
-
-case class DataSetEntry(
-    key: String,
-    value: String,
-    label: String,
-    concatenation: String
-)
-object DataSetEntryBuilder {
-    def fromDataSet(set: DataSet): DataSetEntry = {
-        val prop = set.properties.filter(_.propertyName == "Dataset Name").head.propertyValue
-        DataSetEntry(
-            set.dataSetId.toString,
-            set.dataSetId.toString,
-            prop,
-            prop
-        )
-    }
-}
-
-case class DataAccessRequestDraft(
-    userId: Int,
-    datasets: Seq[DataSetEntry],
-    datasetId: Seq[Int]
-)
-
-case class DataAccessRequest(
-    id: Int,
-    referenceId: String,
-    data: Option[DataAccessRequestData] = None,
-    draft: Option[Boolean] = None,
-    userId: Int,
-    createDate: Option[Long] = None,
-    sortDate: Option[Long] = None,
-    submissionDate: Option[Long] = None,
-    updateDate: Option[Long] = None
-)
-
-case class DataAccessRequestData(
-    referenceId: Option[String] = None,
-    investigator: Option[String] = None,
-    institution: Option[String] = None,
-    department: Option[String] = None,
-    division: Option[String] = None,
-    address1: Option[String] = None,
-    address2: Option[String] = None,
-    city: Option[String] = None,
-    zipCode: Option[String] = None,
-    state: Option[String] = None,
-    country: Option[String] = None,
-    projectTitle: Option[String] = None,
-    checkCollaborator: Option[Boolean] = None,
-    researcher: Option[String] = None,
-    userId: Option[Int] = None,
-    isThePi: Option[String] = None,
-    havePi: Option[String] = None,
-    piEmail: Option[String] = None,
-    profileName: Option[String] = None,
-    pubmedId: Option[String] = None,
-    scientificUrl: Option[String] = None,
-    eraExpiration: Option[Boolean] = None,
-    academicEmail: Option[String] = None,
-    eraAuthorized: Option[Boolean] = None,
-    nihUsername: Option[String] = None,
-    linkedIn: Option[String] = None,
-    orcid: Option[String] = None,
-    researcherGate: Option[String] = None,
-    rus: Option[String] = None,
-    nonTechRus: Option[String] = None,
-    diseases: Option[Boolean] = None,
-    methods: Option[Boolean] = None,
-    controls: Option[Boolean] = None,
-    population: Option[Boolean] = None,
-    other: Option[Boolean] = None,
-    otherText: Option[String] = None,
-    ontologies: Option[Seq[OntologyEntry]] = None,
-    forProfit: Option[Boolean] = None,
-    oneGender: Option[Boolean] = None,
-    gender: Option[String] = None,
-    pediatric: Option[Boolean] = None,
-    illegalBehavior: Option[Boolean] = None,
-    addiction: Option[Boolean] = None,
-    sexualDiseases: Option[Boolean] = None,
-    stigmatizedDiseases: Option[Boolean] = None,
-    vulnerablePopulation: Option[Boolean] = None,
-    populationMigration: Option[Boolean] = None,
-    psychiatricTraits: Option[Boolean] = None,
-    notHealth: Option[Boolean] = None,
-    hmb: Option[Boolean] = None,
-    status: Option[String] = None,
-    poa: Option[Boolean] = None,
-    datasets: Option[Seq[DataSetEntry]] = None,
-    darCode: Option[String] = None,
-    partialDarCode: Option[String] = None,
-    restriction: Option[Any] = None,
-    validRestriction: Option[Boolean] = None,
-    translatedUseRestriction: Option[String] = None,
-    createDate: Option[Long] = None,
-    sortDate: Option[Long] = None,
-    datasetIds: Option[Seq[Int]] = None,
-    datasetDetail: Option[Seq[DataSetDetailEntry]] = None,
-    anvilUse: Option[Boolean] = None,
-    cloudUse: Option[Boolean] = None,
-    localUse: Option[Boolean] = None,
-    cloudProvider: Option[String] = None,
-    cloudProviderType: Option[String] = None,
-    cloudProviderDescription: Option[String] = None,
-    geneticStudiesOnly: Option[Boolean] = None,
-    irb: Option[Boolean] = None,
-    irbDocumentLocation: Option[String] = None,
-    irbDocumentName: Option[String] = None,
-    irbProtocolExpiration: Option[String] = None,
-    itDirector: Option[String] = None,
-    signingOfficial: Option[String] = None,
-    publication: Option[Boolean] = None,
-    collaboration: Option[Boolean] = None,
-    collaborationLetterLocation: Option[String] = None,
-    collaborationLetterName: Option[String] = None,
-    forensicActivities: Option[Boolean] = None,
-    sharingDistribution: Option[Boolean] = None,
-    labCollaborators: Option[Seq[Collaborator]] = None,
-    internalCollaborators: Option[Seq[Collaborator]] = None,
-    externalCollaborators: Option[Seq[Collaborator]] = None,
-    dsAcknowledgement: Option[Boolean] = None,
-    gsoAcknowledgement: Option[Boolean] = None,
-    pubAcknowledgement: Option[Boolean] = None
-)
-
-case class OntologyEntry(
-    id: Option[String],
-    label: Option[String],
-    definition: Option[String],
-    synonyms: Option[Seq[String]]
-)
-
-case class DataSetDetailEntry(
-    datasetId: Option[String],
-    name: Option[String],
-    objectId: Option[String]
-)
-
-case class Collaborator(
-    approverStatus: Option[Boolean],
-    email: Option[String],
-    eraCommonsId: Option[String],
-    name: Option[String],
-    title: Option[String],
-    uuid: Option[String]
-)
-
-case class FireCloudProfile(
-    firstName: Option[String] = None,
-    lastName: Option[String] = None,
-    title: Option[String] = None,
-    contactEmail: Option[String] = None,
-    institute: Option[String] = None,
-    institutionalProgram: Option[String] = None,
-    programLocationCity: Option[String] = None,
-    programLocationState: Option[String] = None,
-    programLocationCountry: Option[String] = None,
-    pi: Option[String] = None,
-    nonProfitStatus: Option[String] = None
-)
-
-case class NihUserAccount(
-    linkedNihUsername: String,
-    datasetPermissions: Option[Seq[String]] = None,
-    linkExpireTime: String,
-    status: Boolean
-)
-
-case class NihVerify(
-    statusCode: Option[String]
-)
-
-case class Dac(
-    dacId: Option[Int],
-    name: Option[String],
-    description: Option[String],
-    createDate: Option[Long],
-    updateDate: Option[Long],
-    chairpersons: Option[User],
-    members: Option[User],
-    electionIds: Option[Seq[Int]]
-)
-
-case class DataAccessRequestManage(
-    referenceId: Option[String],
-    logged: Option[String],
-    alreadyVoted: Option[Boolean],
-    isReminderSent: Option[Boolean],
-    isFinalVote: Option[Boolean],
-    voteId: Option[Int],
-    totalVotes: Option[Int],
-    votesLogged: Option[Int],
-    rpElectionId: Option[Int],
-    rpVoteId: Option[Int],
-    consentGroupName: Option[String],
-    dac: Option[Dac],
-    electionStatus: Option[String],
-    status: Option[String],
-    rus: Option[String],
-    dataRequestId: Option[String],
-    projectTitle: Option[String],
-    frontEndId: Option[String],
-    electionId: Option[Int],
-    createDate: Option[Long],
-    sortDate: Option[Long],
-    electionVote: Option[Boolean],
-    isCanceled: Option[Boolean],
-    needsApproval: Option[Boolean],
-    dataSetElectionResult: Option[String],
-    datasetId: Option[Int],
-    dacId: Option[Int],
-    errors: Option[Seq[String]],
-    ownerUser: Option[User]
-)
 
 object JsonProtocols extends DefaultJsonProtocol {
     implicit val researcherPropertyFormat = jsonFormat4(ResearcherProperty)
@@ -335,7 +24,7 @@ object JsonProtocols extends DefaultJsonProtocol {
     implicit val collaboratorFormat = jsonFormat6(Collaborator)
     implicit val dataSetDetailFormat = jsonFormat3(DataSetDetailEntry)
     implicit val ontologyEntryFormat = jsonFormat4(OntologyEntry)
-    implicit val userFormat = jsonFormat9(User)
+    implicit val userFormat = jsonFormat10(User)
     implicit val dataSetFormat = jsonFormat16(DataSet)
     implicit val dataAccessRequestDraftFormat = jsonFormat3(DataAccessRequestDraft)
     implicit val fireCloudProfileFormat = jsonFormat11(FireCloudProfile)
