+++ automation/src/test/scala/org/broadinstitute/dsp/consent/models/WhiteListModels.scala
@@ -1,0 +1,14 @@
+package org.broadinstitute.dsp.consent.models
+
+object WhiteListModels {
+    case class WhiteListEntry(
+        organization: String,
+        commonsId: String,
+        name: String,
+        email: String,
+        signingOfficialName: String,
+        signingOfficialEmail: String,
+        itDirectorName: String,
+        itDirectorEmail: String
+    )
+}
